Initial commit with crate-docs-theme
This is the contents of crate/crate-docs-theme as of commit
a646cb5843fa60d326741153c19197642e138dab.

We start from this point as the theme is attractive and has
a top navbar separate from the sidebar, which is how we want
to organize all of the Nengo docs pages. We have a lot of concerns
that differ from Crate, however, so we chose to fork the theme
and develop it separately. This is also not necessarily the
permanent Nengo theme moving forward, but a useful intermediate
solution for now.

We are forking this as per the terms of the Apache version 2.0
license, which allow derivative works without having to contribute
the changes back to the original copyright holder.