Test version dropdown in feature test page
The deeply nested link was added because the most complicated
logic in the template is making sure that nested links go to
their correct locations in the other version of the docs.

This also fixes a small bug in the link to the main page when
on pages nested in a folder.