Redesign to match other ABR sites
FusionBox provided the starting point. We added some additional
features, like ScrollSpy, a sticky sidebar, and more admonition
styles.

This commit also switches to using Nengo Bones to manage
meta files in the repository, which has the intended side effect
of using the Nengo license, not the Apache license, as none of the
files in this repository have any crate.io code anymore.

Add notebook test.