Ignore local redirect page
We ignore the redirect link because Sphinx started
checking local links in 3.2 (previously they were
always ignored). But the linkcheck doesn't work, because
Sphinx expects the links to be relative to the source
directory (e.g. docs/redirect/to/nested-page.html).
Because we're generating the pages dynamically, they
only exist in the build directory
(e.g. docs/_build/redirect...).